pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-agent-poc"
version = "1.0.0"
description = "Agentic knowledge extraction from research artifacts"
requires-python = ">=3.10"
readme = "README.md"
authors = [
    {name = "Event Agent Team"}
]

dependencies = [
    # Agent Framework (preview)
    "agent-framework-azure-ai>=0.1.0.dev0",
    
    # Core dependencies
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # API
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.23.0",
    
    # Azure & Identity
    "azure-identity>=1.14.0",
    "azure-ai-evaluation>=0.1.0",
    
    # Document processing
    "pdfplumber>=0.10.0",
    "PyPDF2>=3.0.0",
    "pymupdf>=1.23.0",
    
    # Utilities
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
